YP-019: Error handling fix

+++ api/controllers/photoAlbum.ts
@@ -1,3 +1,4 @@
+import { isValidObjectId } from "mongoose";
 import PhotoAlbumModel from "../models/photoAlbum";
 
 export const createPhotoAlbum = async (req: any, res: any) => {
@@ -26,6 +27,9 @@ export const createPhotoAlbum = async (req: any, res: any) => {
 export const getPhotoAlbums = async (req: any, res: any) => {
   try {
     if (req.params.id) {
+      if (!isValidObjectId(req.params.id)) {
+        return res.status(400).json({ error: "URL inválida" });
+      }
       const photoAlbum = await PhotoAlbumModel.findById(req.params.id);
       if (!photoAlbum) {
         return res.status(404).json({ error: "Álbum de fotos no encontrado" });
@@ -46,6 +50,9 @@ export const deletePhotoAlbum = async (req: any, res: any) => {
     const photoAlbum = await PhotoAlbumModel.findById(photoAlbumId);
     if (!photoAlbum) {
       return res.status(404).json({ error: "Álbum de fotos no encontrado" });
+    }
+    if (!isValidObjectId(req.params.id)) {
+      return res.status(400).json({ error: "Id inválida" });
     }
     await PhotoAlbumModel.deleteOne({ _id: photoAlbumId });
     res.status(200).json({ message: "Álbum de fotos eliminado exitosamente" });
@@ -68,6 +75,10 @@ export const updatePhotoAlbum = async (req: any, res: any) => {
     const photoAlbum = await PhotoAlbumModel.findById(id);
     if (!photoAlbum) {
       return res.status(404).json({ error: "Álbum de fotos no encontrado" });
+    }
+
+    if (!isValidObjectId(id)) {
+      return res.status(400).json({ error: "Id inválida" });
     }
 
     photoAlbum.photos = photos;
